+++ cnd/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Change log level configuration format from capitalised (e.g. "Debug") to lowercase (e.g. "debug").
+
 ## [cnd-0.9.0] - 2020-10-12
 
 ### Added